portfolio/requirements.txt
django-crispy-forms
django
django-widget-tweaks
djangorestframework
requests
Jinja2